requirements.txt
flask
flask-socketio
flask-cors
eventlet
psycopg2-binary
